Automatic Hirst spot painter
Please download the hirst.jpeg image along with the original code and remember to put them together in the same directory. Please ensure that your python console has the module "colorgram.py" installed if not please run "pip3 install colorgram.py".
Hirst is an artist who drew several of these types of paintings and had fetched millions of dollars from it. This project is an attempt to mimic those paintings and recreate them using turtle.